Converted posts/_ tags/_ and blog/_ into _.vue
