--- src/content/tags.yml
@@ -1,4 +1,0 @@
-tags:
-  - title: Portfolio
-    slug: portfolio
-    description:
