--- src/pages/portfolio.vue
@@ -1,7 +1,0 @@
-<script>
-export default {
-  asyncData({ redirect }) {
-    return redirect('/tag/portfolio');
-  }
-};
-</script>
